tests/backend/stage2/test_unit_image_generation.py
import pytest
from unittest.mock import Mock, patch, MagicMock
import io
from PIL import Image


class TestImageGenerationUnit:
    
    @pytest.fixture
    def sample_scene_data(self):
        return {
            "scene_id": "scene_001",
            "order": 1,
            "description": "清晨的城市街道,阳光洒在高楼之间",
            "composition": "远景,俯视角度",
            "characters": ["char_001"],
            "narration": "这是一个平凡的早晨。"
        }
    
    @pytest.fixture
    def sample_character_data(self):
        return {
            "id": "char_001",
            "name": "张三",
            "description": "年轻男子,黑色短发,穿着现代休闲装",
            "personality": "勇敢、善良"
        }
    
    @pytest.fixture
    def mock_image(self):
        img = Image.new('RGB', (1920, 1080), color='blue')
        return img
    
    def test_build_prompt_from_scene_description(self, sample_scene_data, sample_character_data):
        with patch('services.image_generation.prompt_builder.build') as mock_build:
            mock_build.return_value = "清晨的城市街道,阳光洒在高楼之间,一位黑色短发的年轻男子穿着休闲装行走,远景俯视角度,动漫风格,高质量,4K分辨率"
            
            prompt = build_prompt(sample_scene_data, [sample_character_data], style="anime")
            
            assert "清晨" in prompt
            assert "城市街道" in prompt
            assert "黑色短发" in prompt
            assert "动漫风格" in prompt
            mock_build.assert_called_once()
    
    def test_build_prompt_with_multiple_characters(self, sample_scene_data):
        characters = [
            {"id": "char_001", "name": "张三", "description": "年轻男子,黑色短发"},
            {"id": "char_002", "name": "李四", "description": "中年男子,棕色头发"}
        ]
        sample_scene_data["characters"] = ["char_001", "char_002"]
        
        with patch('services.image_generation.prompt_builder.build') as mock_build:
            mock_build.return_value = "场景描述,包含两个角色"
            
            prompt = build_prompt(sample_scene_data, characters)
            
            mock_build.assert_called_once()
            call_args = mock_build.call_args[0][0]
            assert len(call_args["characters"]) == 2
    
    def test_build_prompt_with_style_tags(self, sample_scene_data, sample_character_data):
        styles = ["anime", "realistic", "watercolor"]
        
        for style in styles:
            with patch('services.image_generation.prompt_builder.build') as mock_build:
                mock_build.return_value = f"场景描述,{style}风格"
                
                prompt = build_prompt(sample_scene_data, [sample_character_data], style=style)
                
                assert style in prompt.lower() or "风格" in prompt
    
    @patch('services.image_generation.stable_diffusion.generate')
    def test_generate_image_from_prompt(self, mock_generate, mock_image):
        mock_generate.return_value = mock_image
        
        image = generate_image_from_text(
            prompt="测试提示词",
            width=1920,
            height=1080,
            model="stable-diffusion-xl"
        )
        
        assert image is not None
        assert image.size == (1920, 1080)
        mock_generate.assert_called_once()
    
    def test_generate_image_with_seed(self, mock_image):
        with patch('services.image_generation.stable_diffusion.generate') as mock_generate:
            mock_generate.return_value = mock_image
            
            seed = 123456
            image = generate_image_from_text(
                prompt="测试提示词",
                seed=seed
            )
            
            call_kwargs = mock_generate.call_args[1]
            assert call_kwargs.get("seed") == seed
    
    def test_generate_image_with_steps_parameter(self, mock_image):
        with patch('services.image_generation.stable_diffusion.generate') as mock_generate:
            mock_generate.return_value = mock_image
            
            steps = 30
            image = generate_image_from_text(
                prompt="测试提示词",
                steps=steps
            )
            
            call_kwargs = mock_generate.call_args[1]
            assert call_kwargs.get("steps") == steps
    
    def test_save_image_to_storage(self, mock_image):
        with patch('services.image_generation.storage_client.upload') as mock_upload:
            mock_upload.return_value = "https://cdn.qiniu.com/xxx/scene_001.png"
            
            url = save_image(mock_image, "scene_001.png")
            
            assert url.startswith("https://")
            assert "scene_001.png" in url
            mock_upload.assert_called_once()
    
    def test_generate_character_reference_image(self, sample_character_data, mock_image):
        with patch('services.image_generation.stable_diffusion.generate') as mock_generate:
            mock_generate.return_value = mock_image
            
            ref_image = generate_character_reference(sample_character_data)
            
            assert ref_image is not None
            assert ref_image.size == (1920, 1080)
    
    def test_apply_controlnet_for_character_consistency(self, mock_image):
        reference_image = mock_image
        scene_prompt = "新场景描述"
        
        with patch('services.image_generation.controlnet.apply') as mock_apply:
            mock_apply.return_value = mock_image
            
            result_image = apply_character_consistency(
                reference_image=reference_image,
                scene_prompt=scene_prompt
            )
            
            assert result_image is not None
            mock_apply.assert_called_once()
    
    def test_image_quality_validation(self, mock_image):
        result = validate_image_quality(mock_image, min_width=1920, min_height=1080)
        
        assert result is True
    
    def test_image_quality_validation_failure(self):
        small_image = Image.new('RGB', (800, 600), color='blue')
        
        result = validate_image_quality(small_image, min_width=1920, min_height=1080)
        
        assert result is False
    
    def test_convert_image_format(self, mock_image):
        with patch('PIL.Image.Image.save') as mock_save:
            convert_image_format(mock_image, format='PNG')
            
            mock_save.assert_called_once()
            call_kwargs = mock_save.call_args[1]
            assert call_kwargs.get('format') == 'PNG'
    
    def test_resize_image_maintain_aspect_ratio(self, mock_image):
        resized = resize_image(mock_image, target_width=1280, maintain_aspect=True)
        
        assert resized.width == 1280
        assert resized.height == int(1080 * (1280 / 1920))
    
    def test_cache_generation_parameters(self, sample_scene_data):
        params = {
            "model": "stable-diffusion-xl",
            "seed": 123456,
            "steps": 30,
            "prompt": "测试提示词"
        }
        
        with patch('services.image_generation.cache.set') as mock_cache_set:
            cache_generation_params(sample_scene_data["scene_id"], params)
            
            mock_cache_set.assert_called_once()
            assert mock_cache_set.call_args[0][0] == f"gen_params:{sample_scene_data['scene_id']}"
    
    def test_retrieve_cached_parameters(self, sample_scene_data):
        cached_params = {
            "model": "stable-diffusion-xl",
            "seed": 123456
        }
        
        with patch('services.image_generation.cache.get') as mock_cache_get:
            mock_cache_get.return_value = cached_params
            
            result = get_cached_generation_params(sample_scene_data["scene_id"])
            
            assert result == cached_params
            mock_cache_get.assert_called_once()
    
    def test_fallback_to_dalle_on_sd_failure(self):
        with patch('services.image_generation.stable_diffusion.generate') as mock_sd:
            with patch('services.image_generation.dalle.generate') as mock_dalle:
                mock_sd.side_effect = Exception("SD服务不可用")
                mock_dalle.return_value = Image.new('RGB', (1920, 1080))
                
                image = generate_image_with_fallback(prompt="测试提示词")
                
                assert image is not None
                mock_sd.assert_called_once()
                mock_dalle.assert_called_once()
    
    def test_negative_prompt_handling(self):
        positive_prompt = "美丽的风景"
        negative_prompt = "低质量,模糊,变形"
        
        with patch('services.image_generation.stable_diffusion.generate') as mock_generate:
            mock_generate.return_value = Image.new('RGB', (1920, 1080))
            
            image = generate_image_from_text(
                prompt=positive_prompt,
                negative_prompt=negative_prompt
            )
            
            call_kwargs = mock_generate.call_args[1]
            assert "negative_prompt" in call_kwargs
    
    def test_batch_image_generation(self, sample_scene_data):
        scenes = [sample_scene_data.copy() for _ in range(5)]
        for i, scene in enumerate(scenes):
            scene["scene_id"] = f"scene_{i:03d}"
        
        with patch('services.image_generation.generate_image') as mock_generate:
            mock_generate.return_value = Image.new('RGB', (1920, 1080))
            
            results = generate_images_batch(scenes)
            
            assert len(results) == 5
            assert mock_generate.call_count == 5
    
    def test_lora_model_application(self):
        with patch('services.image_generation.stable_diffusion.generate') as mock_generate:
            mock_generate.return_value = Image.new('RGB', (1920, 1080))
            
            image = generate_image_from_text(
                prompt="测试提示词",
                lora_model="anime_style_v2",
                lora_weight=0.8
            )
            
            call_kwargs = mock_generate.call_args[1]
            assert "lora_model" in call_kwargs
            assert call_kwargs["lora_weight"] == 0.8


def build_prompt(scene_data, characters, style="anime"):
    pass

def generate_image_from_text(prompt, width=1920, height=1080, model="stable-diffusion-xl", seed=None, steps=30, negative_prompt=None, lora_model=None, lora_weight=None):
    pass

def save_image(image, filename):
    pass

def generate_character_reference(character_data):
    pass

def apply_character_consistency(reference_image, scene_prompt):
    pass

def validate_image_quality(image, min_width, min_height):
    pass

def convert_image_format(image, format):
    pass

def resize_image(image, target_width, maintain_aspect=True):
    pass

def cache_generation_params(scene_id, params):
    pass

def get_cached_generation_params(scene_id):
    pass

def generate_image_with_fallback(prompt):
    pass

def generate_images_batch(scenes):
    pass
